raspplayer_backend/requirements.txt
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
btrfsutil==5.12.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
distlib==0.3.1
distro==1.5.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
Flask==1.1.2
Flask-JSON==0.3.4
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gufw==21.4.0
html5lib==1.1
idna==3.1
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.8
keyutils==0.6
LibAppArmor==3.0.1
lightdm-gtk-greeter-settings==1.2.2
mariadb==1.0.7
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.2
menulibre==2.2.1
Mopidy==3.1.1
Mopidy-MPD==3.1.0
msgpack==1.0.2
mugshot==0.4.3
npyscreen==4.10.5
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.2
pep517==0.10.0
pexpect==4.8.0
ply==3.11
progress==1.5
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.20.0
pycparser==2.20
PyGObject==3.40.1
PyJWT==2.1.0
pykka==3.0.1
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
python-distutils-extra==2.39
python-mpd2==3.0.4
pyxdg==0.27
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
six==1.15.0
SQLAlchemy==1.4.10
team==1.0
toml==0.10.2
tornado==6.1
typing==3.7.4.3
udiskie==2.3.3
ufw==0.36
urllib3==1.26.5
webencodings==0.5.1
Werkzeug==1.0.1
